[Refactor] : CropImage -> AD_CropImage 이동
Git : #51
