+++ Projects/UI/AD_UI/Sources/Views/MakeADViews/Step3_SeparatingCharacterView/SeparatingCharacterView.swift
@@ -7,10 +7,10 @@
 //
 
 import SwiftUI
-import AD_MaskingImage
 import AD_Feature
 import AD_Utils
 import ComposableArchitecture
+import AD_MaskingImage
 
 struct SeparatingCharacterView: ADUI {
   typealias MyFeature = SeparatingCharacterFeature
