+++ Projects/UI/AD_UI_Helper/AD_CropImage/Resources/SampleDrawing.xcassets/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
