+++ Projects/UI/AD_UI_Helper/AD_ModifyJoints/Project.swift
@@ -1,18 +1,22 @@
+//
+//  Project.swift
+//  Config
+//
+//  Created by minii on 2023/08/23.
+//
+
 import ProjectDescription
 import ProjectEnvironment
 
 let project = Project.makeModule(
-  myModule: .AD_UI,
+  myModule: .AD_ModifyJoints,
   platform: .iOS,
 //  product: .staticFramework,
   product: .framework,
   dependencies: [
-    .AD_CropImage,
-    .AD_MaskingImage,
-    .AD_ModifyJoints,
-    .AD_Feature,
-    .AD_Utils,
-    .ComposableArchitecture
+    .AD_Utils
   ],
+  resources: ["Resources/**"],
   withTest: false
 )
+
